set up relation the between diary and user. show a screen depending o…

+++ resources/views/diaries/index.blade.php
@@ -15,17 +15,19 @@
         <body>
             <div class='diaries'>
                 @foreach ($diaries as $diary)
-                    <div class='diary'>
-                        <h2 class='title'>{{$diary->title}}</h2>
-                        <p class='content'>{{$diary->content}}</p>
-                        <img src="{{ asset($diary->photo) }}" width="100px">
-                        <a href="/diaries/{{$diary->id}}">see more</a>
-                        <form action="/diaries/{{ $diary->id }}" id="form_{{ $diary->id }}" method="post">
-                            @csrf
-                            @method('DELETE')
-                            <button type="button" onclick="deleteDiary({{ $diary->id }})">Delete</button> 
-                        </form>
-                    </div>
+                    @if($diary->user_id === $user->id)
+                        <div class='diary'>
+                            <h2 class='title'>{{$diary->title}}</h2>
+                            <p class='content'>{{$diary->content}}</p>
+                            <img src="{{ asset($diary->photo) }}" width="100px">
+                            <a href="/diaries/{{$diary->id}}">see more</a>
+                            <form action="/diaries/{{ $diary->id }}" id="form_{{ $diary->id }}" method="post">
+                                @csrf
+                                @method('DELETE')
+                                <button type="button" onclick="deleteDiary({{ $diary->id }})">Delete</button> 
+                            </form>
+                        </div>
+                    @endif
                 @endforeach
             </div>
             <a href="/diaries/create">
